Allow for fractional entry for timesheet hours, change some response messages and payroll notification update during invoice generation.

+++ Models/AppModels/TimeSheet.cs
@@ -8,7 +8,7 @@ namespace TimesheetBE.Models.AppModels
         public bool IsApproved { get; set; }
         public Guid EmployeeInformationId { get; set; }
         public EmployeeInformation EmployeeInformation { get; set; }
-        public int Hours { get; set; }
+        public double Hours { get; set; }
         public int? StatusId { get; set; }
         public Status Status { get; set; }
         public string? RejectionReason { get; set; }
